Create anonymous objects as dynamically created classes.
Turn anonymous objects into instances of a dynamically created class. References to local variables outside the anonymous object
block (uprefs) are turned into instance variables in the new class and initialized from the current value. This is the model that List
uses to create its iterator.

Reference to these upref variables are now generated as messages instead of dot references.

Limitations of this approach are that methods are static and can't be rewritten, references to other fields in the anonymous object
aren't handled and the current code can't handle nested anonymous objects.

The main goal of this is to get iterator working for List.