Add "NativeImpl" meta so we can mark things are implemented natively.
Since we use categories to add functionality to existing classes (e.g. NSString) Objective-C will get upset
if we redefine properties that aren't exactly the same so we can use NativeImpl to keep from generating those
properties (and methods?).

Maybe there was an existing meta for this but I didn't see anything appropriate.