Handle explicit, user defined, setters by wrapping then in a generated setter.
Objective-C requires that setters have a type of void while Haxe requires that a setter return a type
of their properties values. To make this work generate wrapper setters that call the user defined setter
and ignore the value.

This means that we can't assume that a setter has a value, which should be legal in Haxe. Of course
all the IOS library setters don't return values so that wasn't very safe anyway.