Support  native files (non-haze) in the project by adding them using -objc-native.
Support adding non-haxe files to the generated project via the “-objc-native” flag. The argument is either
a single file or a directory relative to the objc output directory.

Files with an extension of “.m” or “.c” will be added to the PBXBuildFile section of the generated Xcode project.

Since the files are assumed to be in the generated output directory some external means will be needed to get them there.

Maybe a bit of a hack but it lets me add Objective-C and C code that hasn’t been converted to Haxe.