Create a method to peek into an expression and get a type.
It seems that some times the result of an expression loses the identity of a parasitized type
resulting in only the base type (Int for Null<Int>) so create a function to find the type of the
expression based on the text_expr itself. Seems mostly to be needed for TCall.