Better static handling of externs and a hack for external statics that aren’t really part of a class.
Use nativeImpl meta data to indicate that a static method can be called without being prefixed by its class.
This is needed by the CGGeometry refs which are defined in CoreGraphics. Haxe doesn’t seem to have a
concept of a plain, C-style, extern class.

Also reset generating_c_call as soon as we’ve determined we’re using it so it doesn’t pollute other generation.